requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
newsapi-python==0.2.7
jinja2==3.1.2
googletrans==3.1.0a0
flask==3.0.2
flask-sqlalchemy==3.1.1
gunicorn==21.2.0 